+++ src/filesystem/create_file.rs
@@ -16,6 +16,6 @@ pub fn create_file(path: &str, content: &str, executable: FileMode) -> Result<()
     if executable.is_executable() {
         set_executable(file_path)?;
     }
-    println!("installed {path}");
+    println!("created {path}");
     Ok(())
 }